Version 1.1

+++ src/main/kotlin/de/redgames/somebot/SomeBot.kt
@@ -3,24 +3,26 @@ package de.redgames.somebot
 import net.dv8tion.jda.api.JDA
 import net.dv8tion.jda.api.OnlineStatus
 import net.dv8tion.jda.api.entities.Activity
-import net.dv8tion.jda.api.events.message.guild.GuildMessageReceivedEvent
+import net.dv8tion.jda.api.events.message.MessageReceivedEvent
 import net.dv8tion.jda.api.hooks.ListenerAdapter
 import java.util.concurrent.TimeUnit
 
 class SomeBot(private val jda: JDA) {
     fun start() {
         jda.addEventListener(object : ListenerAdapter() {
-            override fun onGuildMessageReceived(event: GuildMessageReceivedEvent) = handleGuildMessage(event)
+            override fun onMessageReceived(event: MessageReceivedEvent) = handleGuildMessage(event)
         })
 
         jda.presence.setPresence(OnlineStatus.ONLINE, Activity.watching("@someone"))
     }
 
-    private fun handleGuildMessage(event: GuildMessageReceivedEvent) {
-        val mentionedUsers = event.message.mentionedUsers
+    private fun handleGuildMessage(event: MessageReceivedEvent) {
+        if (!event.isFromGuild) return
+
+        val mentionedUsers = event.message.mentions.users
 
         if (mentionedUsers.contains(jda.selfUser) && event.member != null) {
-            SomeoneMessage.send(event.channel, event.member!!)
+            SomeoneMessage.send(event.channel.asTextChannel(), event.member!!)
             event.message.delete().queueAfter(100, TimeUnit.MILLISECONDS)
         }
     }
